createtasks.php - prototype with CURL
# README

One.UMD is the branded oneCampus solution for the University of Maryland. This PHP script allows an API account (with basic authentication) to submit a local JSON file via CURL to the Task Management interface for approval.

*NOTE : This script is just a prototype for a more complex batch processing script and can only accept only Task per JSON file. *

### Here is the sample JSON file.  
Copy & Paste the JSON below to the same folder as the script and when prompted for the local file just use the name and extension of the JSON file. _I used the filename : newtask.json_

```JSON

{
    "publisherId": 1361,
    "contactId": 1079,
    "title": "TerpZone",
    "statType": "USER",
    "uniqueKey": "terpzone",
    "description": "The TerpZone welcomes anyone from students and faculty to alumni and visitors to enjoy everything the facility offers, including (8) Brunswick Bowling Lanes, (10) Gold Crown 9' Billiard Tables, Video Arcade Games and a Loun